fixing warnings, refs #3
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@24128 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/netedit/GNEEdge.cpp
@@ -419,7 +419,7 @@ GNEEdge::remakeGNEConnections() {
 
     // first check that there are the same number of NBLanes as GNELanes
     while (myLanes.size() < myNBEdge.getLanes().size()) {
-        GNELane* lane = new GNELane(*this, myLanes.size());
+        GNELane* lane = new GNELane(*this, (int)myLanes.size());
         const NBEdge::Lane& laneAttrs = myNBEdge.getLanes().at(myLanes.size());
         myLanes.push_back(lane);
         lane->incRef("GNEEdge::addLane");
